replaced old writers.py
